# openapi/myQSL-v1.yaml
# This file is auto-generated by scripts/generate-openapi.mts. DO NOT EDIT.
openapi: 3.1.0
info:
  title: myQSL Core API
  version: 1.0.0
  description: Electronic QSL Card System API for Amateur Radio Operators
servers:
  - url: "https://myqsl.app"
    description: Production Server
  - url: "http://localhost:8787"
    description: Local Development Server
paths:
  "/api/v1/qsos":
    get:
      summary: List QSOs
      operationId: listQsos
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          required: false
        - name: call
          in: query
          schema:
            type: string
          required: false
        - name: band
          in: query
          schema:
            type: string
          required: false
        - name: mode
          in: query
          schema:
            type: string
          required: false
        - name: date_from
          in: query
          schema:
            type: string
          required: false
        - name: date_to
          in: query
          schema:
            type: string
          required: false
        - name: include_deleted
          in: query
          schema:
            type: boolean
          required: false
      responses:
        200:
          description: List of QSOs
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/QsoRecord"
                  next_cursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - next_cursor
    post:
      summary: Create QSO
      operationId: createQso
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/QsoInput"
      responses:
        201:
          description: QSO created
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/QsoRecord"
                required:
                  - data
        422:
          "$ref": "#/components/responses/ValidationError"
  "/api/v1/qsos/{id}":
    patch:
      summary: Update QSO
      operationId: patchQso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: "If-Match"
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: object
      responses:
        200:
          description: QSO updated
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/QsoRecord"
                required:
                  - data
        412:
          "$ref": "#/components/responses/PreconditionFailed"
    delete:
      summary: Soft delete QSO
      operationId: deleteQso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: "If-Match"
          in: header
          required: true
          schema:
            type: string
      responses:
        204:
          description: "QSO soft-deleted"
        412:
          "$ref": "#/components/responses/PreconditionFailed"
  "/api/v1/qsos/{id}/restore":
    post:
      summary: "Restore soft-deleted QSO"
      operationId: restoreQso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: QSO restored
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/QsoRecord"
                required:
                  - data
  "/api/v1/stations":
    get:
      summary: List Stations
      operationId: listStations
      responses:
        200:
          description: List of Stations
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/StationRecord"
                required:
                  - data
    post:
      summary: Create Station
      operationId: createStation
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/StationInput"
      responses:
        201:
          description: Station created
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/StationRecord"
                required:
                  - data
  "/api/v1/card-templates":
    get:
      summary: List Card Templates
      operationId: listTemplates
      responses:
        200:
          description: List of Templates
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/CardTemplateRow"
                required:
                  - data
    post:
      summary: Create Card Template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: object
              properties:
                name:
                  type: string
                layout:
                  "$ref": "#/components/schemas/CardTemplate"
              required:
                - name
                - layout
      responses:
        201:
          description: Template created
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/CardTemplateRow"
                required:
                  - data
  "/api/v1/card-templates/{id}/background":
    put:
      summary: Upload Template Background Image
      operationId: uploadTemplateBackground
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          "image/png":
            schema:
              type: string
              format: binary
          "image/jpeg":
            schema:
              type: string
              format: binary
      responses:
        201:
          description: Background uploaded
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      key:
                        type: string
                      etag:
                        type: string
                    required:
                      - key
                      - etag
                required:
                  - data
  "/api/v1/cards":
    get:
      summary: List Cards
      operationId: listCards
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          required: false
      responses:
        200:
          description: List of cards
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/CardRow"
                  next_cursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - next_cursor
    post:
      summary: Create Card Draft
      operationId: createCardDraft
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: object
              properties:
                qso_id:
                  type: integer
                template_id:
                  type: integer
              required:
                - qso_id
                - template_id
      responses:
        201:
          description: Draft created
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/CardRow"
                required:
                  - data
  "/api/v1/cards/{id}/image":
    put:
      summary: Attach Rendered Card Image
      operationId: attachCardImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          "image/png":
            schema:
              type: string
              format: binary
      responses:
        200:
          description: "Image attached, card moved to ready"
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/CardRow"
                required:
                  - data
        409:
          "$ref": "#/components/responses/ConflictError"
  "/api/v1/cards/{id}/publish":
    post:
      summary: Publish Card
      operationId: publishCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Card published
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/CardRow"
                required:
                  - data
        409:
          "$ref": "#/components/responses/ConflictError"
  "/api/v1/cards/{id}/void":
    post:
      summary: Void Published Card
      operationId: voidCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Card voided
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/CardRow"
                required:
                  - data
        409:
          "$ref": "#/components/responses/ConflictError"
  "/api/v1/public/card-lookup":
    post:
      summary: Public Card Lookup
      operationId: publicCardLookup
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: object
              properties:
                call:
                  type: string
                qso_date:
                  type: string
              required:
                - call
                - qso_date
      responses:
        200:
          description: Matching published cards
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/PublicCardSummary"
                required:
                  - data
        429:
          description: Rate limit exceeded
components:
  schemas:
    QsoInput:
      type: object
      properties:
        station_id:
          type: integer
        station_callsign:
          type: string
        call:
          type: string
        qso_date:
          type: string
          pattern: "^\\d{8}$"
        time_on:
          type: string
          pattern: "^\\d{4}(\\d{2})?$"
        band:
          type: string
        mode:
          type: string
        submode:
          type: string
          nullable: true
        freq_mhz:
          type: string
          nullable: true
        rst_sent:
          type: string
          nullable: true
        rst_rcvd:
          type: string
          nullable: true
        gridsquare:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        qth:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        adif_extra:
          type: object
          additionalProperties:
            type: string
      required:
        - station_callsign
        - call
        - qso_date
        - time_on
        - band
        - mode
    QsoRecord:
      type: object
      properties:
        id:
          type: integer
        station_id:
          type: integer
          nullable: true
        station_callsign:
          type: string
        call:
          type: string
        qso_date:
          type: string
        time_on:
          type: string
        band:
          type: string
        mode:
          type: string
        submode:
          type: string
          nullable: true
        freq_mhz:
          type: string
          nullable: true
        rst_sent:
          type: string
          nullable: true
        rst_rcvd:
          type: string
          nullable: true
        gridsquare:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        qth:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        adif_extra_json:
          type: string
        version:
          type: integer
        deleted_at:
          type: integer
          nullable: true
        created_at:
          type: integer
        updated_at:
          type: integer
      required:
        - id
        - station_callsign
        - call
        - qso_date
        - time_on
        - band
        - mode
        - version
    StationInput:
      type: object
      properties:
        callsign:
          type: string
        station_callsign:
          type: string
          nullable: true
        operator_callsign:
          type: string
          nullable: true
        grid_square:
          type: string
          nullable: true
        qth:
          type: string
          nullable: true
        rig:
          type: string
          nullable: true
        antenna:
          type: string
          nullable: true
        power_w:
          type: integer
          nullable: true
        is_default:
          type: boolean
      required:
        - callsign
    StationRecord:
      type: object
      properties:
        id:
          type: integer
        callsign:
          type: string
        station_callsign:
          type: string
          nullable: true
        operator_callsign:
          type: string
          nullable: true
        grid_square:
          type: string
          nullable: true
        qth:
          type: string
          nullable: true
        rig:
          type: string
          nullable: true
        antenna:
          type: string
          nullable: true
        power_w:
          type: integer
          nullable: true
        is_default:
          type: boolean
        version:
          type: integer
        created_at:
          type: integer
        updated_at:
          type: integer
      required:
        - id
        - callsign
        - is_default
        - version
    CardTemplate:
      type: object
      properties:
        schema_version:
          type: integer
          const: 1
        base_width:
          type: integer
        base_height:
          type: integer
        elements:
          type: array
          items:
            type: object
      required:
        - schema_version
        - base_width
        - base_height
        - elements
    CardTemplateRow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        schema_version:
          type: integer
        base_width:
          type: integer
        base_height:
          type: integer
        layout_json:
          type: string
        background_r2_key:
          type: string
          nullable: true
        background_sha256:
          type: string
          nullable: true
        version:
          type: integer
        created_at:
          type: integer
        updated_at:
          type: integer
      required:
        - id
        - name
        - schema_version
        - base_width
        - base_height
        - layout_json
        - version
    CardRow:
      type: object
      properties:
        id:
          type: string
        public_id:
          type: string
        qso_id:
          type: integer
        template_id:
          type: integer
        status:
          type: string
          enum:
            - draft
            - ready
            - published
            - void
        qso_snapshot_json:
          type: string
        template_snapshot_json:
          type: string
        image_r2_key:
          type: string
          nullable: true
        image_sha256:
          type: string
          nullable: true
        lookup_call:
          type: string
          nullable: true
        lookup_qso_date:
          type: string
          nullable: true
        created_at:
          type: integer
        updated_at:
          type: integer
      required:
        - id
        - public_id
        - qso_id
        - template_id
        - status
    PublicCardSummary:
      type: object
      properties:
        public_id:
          type: string
        call:
          type: string
        qso_date:
          type: string
        created_at:
          type: integer
      required:
        - public_id
        - call
        - qso_date
        - created_at
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
        - type
        - title
        - status
        - instance
  responses:
    ValidationError:
      description: Validation error
      content:
        "application/problem+json":
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    PreconditionFailed:
      description: Optimistic concurrency check failed
      content:
        "application/problem+json":
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    ConflictError:
      description: State transition conflict
      content:
        "application/problem+json":
          schema:
            "$ref": "#/components/schemas/ProblemDetails"